# geoai-service/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GeoAI Tool Adapter Service",
    "description": "Controlled, typed tool surface between an AI agent and the GeoSphere platform's PostGIS database, MinIO object storage, and internal APIs. The agent never receives credentials or internal URLs \u2014 every call here is validated, rate-limited, cached, and logged.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/geoai/tools/definitions": {
      "get": {
        "tags": [
          "ai-tools"
        ],
        "summary": "Tool Definitions",
        "operationId": "tool_definitions_geoai_tools_definitions_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Tool Definitions Geoai Tools Definitions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/geoai/tools/execute": {
      "post": {
        "tags": [
          "ai-tools"
        ],
        "summary": "Execute Tool",
        "operationId": "execute_tool_geoai_tools_execute_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolExecuteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolExecuteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ToolExecuteRequest": {
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "reverse_geocode",
              "search_place",
              "find_nearest_place",
              "query_spatial_layer",
              "get_route"
            ],
            "title": "Name"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Caller-supplied conversation/session id, for audit logging."
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ToolExecuteRequest"
      },
      "ToolExecuteResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "title": "Status",
            "default": "success"
          },
          "tool": {
            "type": "string",
            "enum": [
              "reverse_geocode",
              "search_place",
              "find_nearest_place",
              "query_spatial_layer",
              "get_route"
            ],
            "title": "Tool"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "tool"
        ],
        "title": "ToolExecuteResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}